add render to docs
