+++ lang/docs/src/render/mod.rs
@@ -1,0 +1,3 @@
+mod html;
+
+pub use html::*;
