+++ lang/docs/src/util.rs
@@ -1,9 +1,36 @@
 use std::fs;
+
 use std::path::{Path, PathBuf};
 
 use driver::CSS_PATH;
 
-use printer::get_target_path;
+pub fn get_target_path(path: &Path) -> PathBuf {
+    let cwd = std::env::current_dir().expect("Failed to get current working directory");
+    let cwd_name = cwd.file_name().expect("Failed to get current working directory name");
+
+    let mut components = path.components().peekable();
+    let mut new_path = PathBuf::new();
+
+    for component in components.by_ref() {
+        new_path.push(component);
+        if component.as_os_str() == cwd_name {
+            new_path.push("target_pol/docs");
+            break;
+        }
+    }
+
+    for component in components {
+        new_path.push(component);
+    }
+
+    let stem = new_path.file_stem().map(|s| s.to_os_string());
+    if let Some(stem) = stem {
+        new_path.set_file_name(stem);
+        new_path.set_extension("html");
+    }
+
+    new_path
+}
 
 pub fn get_files(folders: Vec<&Path>) -> Vec<(PathBuf, PathBuf)> {
     let mut pol_files = Vec::new();
