+++ lang/printer/src/render/mod.rs
@@ -1,7 +1,5 @@
-mod html;
 mod latex;
 mod termcolor;
 
-pub use html::*;
 pub use latex::*;
 pub use termcolor::*;
